+++ resources/views/pages/partials/blocks/card.blade.php
@@ -11,7 +11,8 @@
         ? $imageAlt
         : trim((string) ($media?->alt_text ?: $media?->title ?: $imageCaption ?: $title ?: 'Card image'));
     $imagePosition = $block->cardImagePosition();
-    $showTopImage = $imageSource && $imagePosition === 'top';
+    $imageAlign = $block->cardImageAlign();
+    $showsImage = $imageSource !== null;
     $renderedDescription = app(\App\Support\Formatting\InlineRichTextRenderer::class)->render($description);
     $actionLabel = trim((string) ($block->meta ?? ''));
     $url = $block->cardUrl();
@@ -19,25 +20,36 @@
     $footerBlocks = $block->children;
     $hasFooterBlocks = $footerBlocks->isNotEmpty();
     $showsLegacyAction = ! $hasFooterBlocks && $url !== null && $actionLabel !== '';
+    $imageFigureClasses = ['wb-stack', 'wb-gap-1'];
+
+    if ($imageAlign === 'stretch') {
+        $imageFigureClasses[] = 'wb-w-full';
+    } elseif ($imageAlign === 'center') {
+        $imageFigureClasses[] = 'wb-text-center';
+    } elseif ($imageAlign === 'end') {
+        $imageFigureClasses[] = 'wb-text-right';
+    }
+
+    $imageFigureClass = implode(' ', $imageFigureClasses);
 @endphp
 
 @if ($block->isPromoCard())
     <section class="wb-card wb-promo" data-wb-public-block-type="{{ $block->publicBlockTypeAttribute() }}">
-        @if ($showTopImage)
-            <figure>
-                <img
-                    src="{{ $imageSource }}"
-                    alt="{{ $resolvedImageAlt }}"
-                    @if ($media?->width) width="{{ $media->width }}" @endif
-                    @if ($media?->height) height="{{ $media->height }}" @endif
-                >
-                @if ($imageCaption !== '')
-                    <figcaption>{{ $imageCaption }}</figcaption>
-                @endif
-            </figure>
-        @endif
-
         <div class="wb-card-body wb-promo-copy wb-stack wb-gap-3">
+            @if ($showsImage && in_array($imagePosition, ['top', 'middle'], true))
+                <figure class="{{ $imageFigureClass }}">
+                    <img
+                        src="{{ $imageSource }}"
+                        alt="{{ $resolvedImageAlt }}"
+                        @if ($media?->width) width="{{ $media->width }}" @endif
+                        @if ($media?->height) height="{{ $media->height }}" @endif
+                    >
+                    @if ($imageCaption !== '')
+                        <figcaption>{{ $imageCaption }}</figcaption>
+                    @endif
+                </figure>
+            @endif
+
             @if ($eyebrow !== '')
                 <p class="wb-eyebrow">{{ $eyebrow }}</p>
             @endif
@@ -52,6 +64,20 @@
 
             @if ($description !== '')
                 <p class="wb-promo-text">{!! $renderedDescription !!}</p>
+            @endif
+
+            @if ($showsImage && $imagePosition === 'bottom')
+                <figure class="{{ $imageFigureClass }}">
+                    <img
+                        src="{{ $imageSource }}"
+                        alt="{{ $resolvedImageAlt }}"
+                        @if ($media?->width) width="{{ $media->width }}" @endif
+                        @if ($media?->height) height="{{ $media->height }}" @endif
+                    >
+                    @if ($imageCaption !== '')
+                        <figcaption>{{ $imageCaption }}</figcaption>
+                    @endif
+                </figure>
             @endif
 
             @if ($hasFooterBlocks || $showsLegacyAction)
@@ -69,29 +95,57 @@
     </section>
 @else
     <article class="wb-card" data-wb-public-block-type="{{ $block->publicBlockTypeAttribute() }}">
-        @if ($showTopImage)
-            <figure>
-                <img
-                    src="{{ $imageSource }}"
-                    alt="{{ $resolvedImageAlt }}"
-                    @if ($media?->width) width="{{ $media->width }}" @endif
-                    @if ($media?->height) height="{{ $media->height }}" @endif
-                >
-                @if ($imageCaption !== '')
-                    <figcaption>{{ $imageCaption }}</figcaption>
-                @endif
-            </figure>
-        @endif
-
         @if ($subtitle !== '')
             <div class="wb-card-header">{{ $subtitle }}</div>
         @endif
 
         <div class="wb-card-body wb-stack wb-gap-2">
+            @if ($showsImage && $imagePosition === 'top')
+                <figure class="{{ $imageFigureClass }}">
+                    <img
+                        src="{{ $imageSource }}"
+                        alt="{{ $resolvedImageAlt }}"
+                        @if ($media?->width) width="{{ $media->width }}" @endif
+                        @if ($media?->height) height="{{ $media->height }}" @endif
+                    >
+                    @if ($imageCaption !== '')
+                        <figcaption>{{ $imageCaption }}</figcaption>
+                    @endif
+                </figure>
+            @endif
+
             <strong>{{ $title }}</strong>
+
+            @if ($showsImage && $imagePosition === 'middle')
+                <figure class="{{ $imageFigureClass }}">
+                    <img
+                        src="{{ $imageSource }}"
+                        alt="{{ $resolvedImageAlt }}"
+                        @if ($media?->width) width="{{ $media->width }}" @endif
+                        @if ($media?->height) height="{{ $media->height }}" @endif
+                    >
+                    @if ($imageCaption !== '')
+                        <figcaption>{{ $imageCaption }}</figcaption>
+                    @endif
+                </figure>
+            @endif
 
             @if ($description !== '')
                 <p class="wb-m-0">{!! $renderedDescription !!}</p>
+            @endif
+
+            @if ($showsImage && $imagePosition === 'bottom')
+                <figure class="{{ $imageFigureClass }}">
+                    <img
+                        src="{{ $imageSource }}"
+                        alt="{{ $resolvedImageAlt }}"
+                        @if ($media?->width) width="{{ $media->width }}" @endif
+                        @if ($media?->height) height="{{ $media->height }}" @endif
+                    >
+                    @if ($imageCaption !== '')
+                        <figcaption>{{ $imageCaption }}</figcaption>
+                    @endif
+                </figure>
             @endif
         </div>